Change double value type to float

+++ Assets/Scripts/Models/Cordinates.cs
@@ -4,12 +4,12 @@
 public class Coordinates
 {
     public string name;
-    public double x;
-    public double y;
-    public double z;
-    public double rotX;
-    public double rotY;
-    public double rotZ;
-    public double rotW;
+    public float x;
+    public float y;
+    public float z;
+    public float rotX;
+    public float rotY;
+    public float rotZ;
+    public float rotW;
 }
